% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instamap.R
\name{state_map}
\alias{state_map}
\title{State Map}
\usage{
state_map(df_path, state, county_label, var_label)
}
\arguments{
\item{df_path}{The path of your csv file.}

\item{state}{The US state to be visualized.}

\item{county_label}{The column label where the state counties are in your csv file.}

\item{var_label}{The column label where your variable of choice is in your csv file.}
}
\value{
A US state map with counties as a ggplot plot object.
}
\description{
This function generates a map of a US state
from the ggplot2 and maps packages while displaying
county outlines and the variable of your choice as a color gradient.
}
\examples{
state_map("crimes.csv", "New York", "County", "Homicide Rate")
state_map("weather.csv", "washington", "county", "Precipitation Rate")
}
