% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/instamap.R
\name{us_map}
\alias{us_map}
\title{US Map}
\usage{
us_map(df_path, state_label, var_label)
}
\arguments{
\item{df_path}{The path of your csv file.}

\item{state_label}{The column label where the US states are in your csv file.}

\item{var_label}{The column label where your variable of choice is in your csv file.}
}
\value{
A US map as a ggplot plot object.
}
\description{
This function generates a map of the lower 48 US states
from the ggplot2 and maps packages while displaying the
variable of your choice as a color gradient.
}
\examples{
us_map("crimes.csv", "State", "Homicide Rate")
us_map("weather.csv", "state", "snowfall")
}
